task

+++ app/Http/Controllers/TasksController.php
@@ -13,22 +13,31 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-   public function index()
-{
-    $tasks = Task::all(); // タスク一覧を取得
-
-    // タスク一覧ビューでそれを表示
-    return view('tasks.index', [
-        'tasks' => $tasks,
-    ]);
-}
+    public function index()
+    {
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
+    }
 
       /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
-         // getでmessages/createにアクセスされた場合の「新規登録画面表示処理
+         // getでtasks/createにアクセスされた場合の「新規登録画面表示処理
     public function create()
     {
          $tasks = new Task;
@@ -45,24 +54,21 @@ class TasksController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-     // postでmessages/にアクセスされた場合の「新規登録処理」
+     // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
-          // バリデーション
+           // バリデーション
         $request->validate([
-            'content' => 'required',
-            'status' => 'required|max:10',
+            'content' => 'required|max:255',
         ]);
+        
 
-
-            //タスクを作成
-        $task = new Task;
-         $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+       // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+        ]);
+         // 前のURLへリダイレクトさせる
+        return back();
     }
 
     /**
@@ -71,13 +77,13 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-      // getでmessages/idにアクセスされた場合の「取得表示処理」
+      // getでtasks/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-          // idの値でメッセージを検索して取得
+          // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        // タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -89,13 +95,13 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-         // getでmessages/id/editにアクセスされた場合の「更新画面表示処理」
+         // getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-         // idの値でメッセージを検索して取得
+         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
+        // タスク編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
         ]);
@@ -108,7 +114,7 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-     // putまたはpatchでmessages/idにアクセスされた場合の「更新処理」
+     // putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
     public function update(Request $request, $id)
     {
           $request->validate([
@@ -117,9 +123,9 @@ class TasksController extends Controller
         ]);
 
 
-         // idの値でメッセージを検索して取得
+         // idの値でtaskを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを更新
+        //タスクを更新
            $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -134,15 +140,19 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-      // deleteでmessages/idにアクセスされた場合の「削除処理」
+      // deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-             // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+       // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
+        }
+       // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 }
